ADD call to methods in the main reconciler function
Calling GenerateTcpdumpArgs to get the in one string the arguments that
will be passed to podtracer as a -a option for tcpdump to run. This can
be generalized in the future replacing the pcap package and tcpdump API
by a podtracer api whenever running any tool using podtracer.

Calling GeneratePodtracerArgs put togethers exclusive podtracer args such
as --pod and -n for namespace.

Calling GenerateTcpdumpJob generates a batchv1.Job object prepared to run
the podtracer image with the arguments collected by the CR. That can be also
generalized or be part of a collection of objects to run multiple different
tools using podtracer. It takes the potracer arguments alongside the target
pod name for placing that reference as part of the name and also the target
node name in order to have reachability to that container.