ADD captured network packets output to container logs
Fixed missing Run command for podtracer.
Commented out the -w argument generation in order to have packet on logs.

The intention behind this change is to test the container logging system
as a means to expose the captured packets to be queried by another workload
and rerouted to a central storage to be processed and analyzed.