ADD finalizer logic to CommandJobReconciler
Since commandjob is supposed to be created on the tenant's namespace
but creates dependent objects such as cronjobs on the snoopy-operator
namespace deletion logic needs to be added because garbage collection
is not allowed in this case. This is cross-namespace owner reference
case that is not allowed by design on k8s.

The separation here is for security reasons. Allow the tenant to create
a CRD that can spin up multiple privileged workers that he doesn't
have access because they are running on a forbidden namespace. The
cronjobs and pods generate are NOT supposed to stay on the same NS
where the tenant CR is created.