ADD Debug and Info logging messages
Controller runtime provides logr as an API to its opinionated way
to write logs. It's using zap logger on the backed via the
controller-runtime/log package.

The logger is instantiated in the manager context and passed down
to the reconcile function that will on it's turn pass the ctx
down with the same logger to other called methods.