ADD documentation on how to use the operator
Included how to use the operator with the single snoopy-operator.yaml
manifest.

Removed the iperf sample that had issues with the packet capture in
favour of a simpler example. Bringing the iperf example back on issue #73

Each new example should have its separate folder with a dedicated README
file documeting each tool we support in snoopy-operator.

Signed-off-by: acmenezes <adcmenezes@gmail.com>